% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_na_links.R
\name{describe_na_links}
\alias{describe_na_links}
\title{Explore MAR hypothesis}
\usage{
describe_na_links(var, var_list, df_md)
}
\arguments{
\item{var}{A character vector, one or more variables with missing data.}

\item{var_list}{A character vector, the potential predictors of is.na(var),}

\item{df_md}{A data.frame, the underlying (unimputed) data.frame}
}
\value{
A data.frame, one row per \code{var}, with the number of associated variables and the Odds details.
}
\description{
Missing at Random (MAR) mechanism of missing data exploration, by bivariate analyses.
}
\details{
Exploration of potential mechanism of missingness is mandatory before performing missing data imputation. The two acceptable settings are the Missing At Random (MAR) mechanism and Missing Completely At Random (MCAR). MAR may be assumed under certain conditions. One of them is that missingness of variables is associated with other variables (the aim is to reduce the plausibility of the last mechanism, Missing Not At Random). This function performs bivariate analyses via glm, between \code{is.na(var)}, where \code{var} is the variable of interest containing missing data, and one or more explaining variables, from \code{var_list}.
The function accepts more than one \code{var} to be evaluated, in the \code{var} argument.
}
\examples{
df_md <-
  data.frame(
    shuffle = rnorm(1000)
  )  |>
  dplyr::mutate(
    v1 = c(rep(1, 300), rep(0, 700)),
    v1 = c(rep(1, 250), rep(NA, 50), rep(0, 700)),
    v2 = ifelse(shuffle < - 0.8, 1, 0),
    v3 = c(rep(1, 200), rep(0, 70), rep(1, 30), rep(0, 700)),
    across(c(v1, v2, v3), as.factor)
  )

md_diag <-
  describe_na_links(
    var = "v1",
    var_list = c("v3"),
    df_md = df_md
  )
}
\keyword{data}
\keyword{imputation}
\keyword{missing}
