% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_date.R
\name{extract_date}
\alias{extract_date}
\title{Extract last modified date of file}
\usage{
extract_date(path_file)
}
\arguments{
\item{path_file}{A character string, the path to the file.}
}
\value{
A character string, the last modified date of the file.
}
\description{
Short-hand around \code{file.info()}
}
\details{
The function examines a file located at \code{path_file} and returns the last modified date (as a character string.
}
\examples{
tf <-
  tempfile()

 write.csv2(mtcars, file = tf)

 extract_date(tf)

 file.remove(tf)
}
\keyword{file}
