% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_eff.R
\name{imp_eff}
\alias{imp_eff}
\title{Average effectives after multiple imputation}
\usage{
imp_eff(var, imp_data)
}
\arguments{
\item{var}{A character vector, the variable to count the effectives of.}

\item{imp_data}{The long format \code{data.table} of imputed datasets.}
}
\description{
This function provides averaged effectives for categorical variables after missing data imputation.
}
\details{
You need to provide imputed datasets in a long format \code{data.table}, not containing the initial dataset. For example, with the \code{mice} package, \code{imp_data <- mice::complete(imp, "long", include = FALSE)}.
}
\examples{
## not yet ready

}
\keyword{mice}
\keyword{missing_data}
